pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "leapmotor-api"
version = "1.0.0"
description = "Unofficial Python client for the Leapmotor vehicle cloud API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "markoceri" },
]
keywords = ["leapmotor", "ev", "vehicle", "api", "iot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "requests>=2.32.4",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
image = [
    "Pillow>=10.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/markoceri/leapmotor-api"
Issues = "https://github.com/markoceri/leapmotor-api/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/leapmotor_api"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
